status ofJohn Doty's predict reddit project code when submitted at UCLA
This snapshot contains code that was a prototype/proof of concept of a
reddit.com (a social news aggregation site) data mining and modeling tool.  The
eventual goal is to create a program that can predict what posts will become
popular before they do.  The current state of the modeling component of this
project captures some of the nonlinear behavior or reddit but as of yet is not
predictive.  Work on the predictive aspect of this project is continuing and
currently focusing on using genetic algorithms to learn the the causal
relationships between different objects and properties of reddit.